backend/models/yolo_detector.py
import os
import torch
import numpy as np
from ultralytics import YOLO
from PIL import Image
import cv2

class YOLODetector:
    def __init__(self, model_path=None):
        """
        Initialize the YOLO detector.
        Since we don't have a trained welding defect model, we'll use YOLOv8 
        and simulate defect detection for demonstration purposes.
        """
        # Use YOLOv8 nano model for faster inference
        self.model = YOLO('yolov8n.pt')
        
        # Define defect types for welding inspection
        self.defect_classes = {
            'crack': 0,
            'porosity': 1,
            'slag': 2,
            'inclusion': 3,
            'undercut': 4,
            'burn_through': 5
        }
        
        # Reverse mapping for class names
        self.class_names = {v: k for k, v in self.defect_classes.items()}
        
    def detect_defects(self, image):
        """
        Detect welding defects in the image.
        For demonstration purposes, we'll simulate detection results.
        In a real implementation, this would use a trained model.
        """
        # Convert PIL image to numpy array
        img_array = np.array(image)
        
        # Run YOLOv8 detection (this will detect general objects)
        results = self.model(img_array)
        
        # Since we don't have a trained welding defect model,
        # we'll simulate some defect detections based on the image characteristics
        simulated_detections = self._simulate_defect_detection(image)
        
        return simulated_detections
    
    def _simulate_defect_detection(self, image):
        """
        Simulate defect detection for demonstration purposes.
        In a real implementation, this would be replaced with actual model inference.
        """
        width, height = image.size
        
        # Create some realistic simulated detections
        detections = []
        
        # Simulate finding defects based on image characteristics
        # This is purely for demonstration - real implementation would use trained model
        
        # Simulate crack detection
        if self._should_detect_crack(image):
            detections.append({
                'class': 'crack',
                'confidence': 0.92,
                'bbox': {
                    'x': int(width * 0.3),
                    'y': int(height * 0.25),
                    'width': int(width * 0.15),
                    'height': int(height * 0.08)
                }
            })
        
        # Simulate porosity detection
        if self._should_detect_porosity(image):
            detections.append({
                'class': 'porosity',
                'confidence': 0.87,
                'bbox': {
                    'x': int(width * 0.6),
                    'y': int(height * 0.45),
                    'width': int(width * 0.1),
                    'height': int(height * 0.12)
                }
            })
        
        # Simulate slag inclusion detection
        if self._should_detect_slag(image):
            detections.append({
                'class': 'slag',
                'confidence': 0.78,
                'bbox': {
                    'x': int(width * 0.2),
                    'y': int(height * 0.7),
                    'width': int(width * 0.12),
                    'height': int(height * 0.06)
                }
            })
        
        return detections
    
    def _should_detect_crack(self, image):
        """
        Simple heuristic to determine if we should simulate a crack detection.
        In reality, this would be based on actual model inference.
        """
        # Convert to grayscale for analysis
        gray = image.convert('L')
        img_array = np.array(gray)
        
        # Look for linear features (potential cracks)
        # This is a very simplified approach
        edges = cv2.Canny(img_array, 50, 150)
        lines = cv2.HoughLinesP(edges, 1, np.pi/180, threshold=50, minLineLength=30, maxLineGap=10)
        
        return lines is not None and len(lines) > 5
    
    def _should_detect_porosity(self, image):
        """
        Simple heuristic to determine if we should simulate porosity detection.
        """
        # Convert to grayscale
        gray = image.convert('L')
        img_array = np.array(gray)
        
        # Look for circular features (potential porosity)
        circles = cv2.HoughCircles(img_array, cv2.HOUGH_GRADIENT, 1, 20,
                                  param1=50, param2=30, minRadius=5, maxRadius=50)
        
        return circles is not None and len(circles[0]) > 2
    
    def _should_detect_slag(self, image):
        """
        Simple heuristic to determine if we should simulate slag inclusion detection.
        """
        # Convert to grayscale
        gray = image.convert('L')
        img_array = np.array(gray)
        
        # Look for irregular bright spots (potential slag inclusions)
        _, thresh = cv2.threshold(img_array, 0, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU)
        contours, _ = cv2.findContours(thresh, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
        
        # Filter contours by size and shape
        valid_contours = 0
        for contour in contours:
            area = cv2.contourArea(contour)
            if 100 < area < 5000:  # Reasonable size range
                valid_contours += 1
        
        return valid_contours > 3
